add casting object

+++ BelajarOOP/src/com/juaracoding/main/Ruang2D.java
@@ -2,8 +2,8 @@ package com.juaracoding.main;
 
 public interface Ruang2D {
 
-	public int luas ();
-	public int keliling ();
+	public float luas ();
+	public float keliling ();
 	
 	
 }
